+++ django/monitor/alerts/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.contrib.auth.models import User
+from applications.models import Service, Application
 from django.template.defaultfilters import truncatechars
-
 
 class AbstractAlert(models.Model):
     creation_date = models.DateTimeField(auto_now_add=True, editable=False, help_text="Creation date")
@@ -12,7 +12,7 @@ class AbstractAlert(models.Model):
         blank=False,
         editable=False,
         help_text="Fingerprint provided by prometheus"
-    )
+    )    
 
     class Meta:
         abstract = True
@@ -21,15 +21,7 @@ class AbstractAlert(models.Model):
     def short_json(self):
         return "%s..." % truncatechars(self.json, 70)
 
-
 class GenericAlert(AbstractAlert):
-
-    class Meta:
-        verbose_name = "Unknown alert"
-        verbose_name_plural = "Unknown alerts"
-
-# Create your models here.
-class InstanceDownAlert(AbstractAlert):
 
     STATUS = [
         (0, 'unknown'),
@@ -45,10 +37,51 @@ class InstanceDownAlert(AbstractAlert):
 
     user = models.ForeignKey(
         User,
+        on_delete = models.CASCADE,
+        related_name = "genericalerts",
+        blank = True,
+        null = True,
+    )
+
+    status = models.IntegerField(choices=STATUS)
+    severity = models.IntegerField(choices=SEVERITY)
+
+    startsAt = models.DateTimeField(null=False)
+    endsAt = models.DateTimeField(null=True, blank=True)
+
+    instance = models.TextField(null=True)
+    summary = models.TextField(null=True)
+    description = models.TextField(null=True)
+
+    @property
+    def duration(self):
+        if self.endsAt:
+            return  self.endsAt - self.startsAt
+        return ''
+    
+    class Meta:
+        verbose_name = "Unknown alert"
+        verbose_name_plural = "Unknown alerts"
+
+class ApplicationAlert(AbstractAlert):
+
+    STATUS = [
+        (0, 'unknown'),
+        (1, 'resolved'),
+        (2, 'firing'),
+    ]
+
+    SEVERITY = [
+        (0, 'unknown'),
+        (1, 'warning'),
+        (2, 'critical'),
+    ]
+
+    application = models.ForeignKey(
+        Application,
         null=True,
         on_delete = models.CASCADE,
-        related_name = "alerts",
-        related_query_name = "alert",
+        related_name = "applicationalerts",
     )
 
     status = models.IntegerField(choices=STATUS)
@@ -68,6 +101,49 @@ class InstanceDownAlert(AbstractAlert):
         return ''
 
     class Meta:
-        verbose_name = "Instance down alert"
-        verbose_name_plural = "Instance down alerts"
+        verbose_name = "Application alert"
+        verbose_name_plural = "Application alerts"
+
+
+# Create your models here.
+class InstanceDownAlert(AbstractAlert):
+
+    STATUS = [
+        (0, 'unknown'),
+        (1, 'resolved'),
+        (2, 'firing'),
+    ]
+
+    SEVERITY = [
+        (0, 'unknown'),
+        (1, 'warning'),
+        (2, 'critical'),
+    ]
+
+    service = models.ForeignKey(
+        Service,
+        null=True,
+        on_delete = models.CASCADE,
+        related_name = "instancedownalerts",
+    )
+
+    status = models.IntegerField(choices=STATUS)
+    severity = models.IntegerField(choices=SEVERITY)
+
+    startsAt = models.DateTimeField(null=False)
+    endsAt = models.DateTimeField(null=True, blank=True)
+
+    instance = models.TextField(null=True)
+    summary = models.TextField(null=True)
+    description = models.TextField(null=True)
+
+    @property
+    def duration(self):
+        if self.endsAt:
+            return  self.endsAt - self.startsAt
+        return ''
+
+    class Meta:
+        verbose_name = "Service down alert"
+        verbose_name_plural = "Service down alerts"