+++ django/monitor/clients/templatetags/remove_protocol.py
@@ -1,0 +1,7 @@
+from django import template
+
+register = template.Library()
+
+@register.filter
+def remove_protocol(value):
+    return value.replace("http://","").replace("https://","")