+++ django/monitor/home/views.py
@@ -1,6 +1,30 @@
 from django.shortcuts import render
 from django.http import HttpResponse
+from django.conf import settings
+
+from rest_framework.authtoken.models import Token
+from django.shortcuts import render
+from clients.models import Server
+import datetime
+from django.utils import timezone
 
 # Create your views here.
 def index(request):
-    return HttpResponse("Hello, world. You're at the monitor index.")
+
+    token = None
+    is_staff = False
+    servers = []
+
+    if not request.user.is_anonymous:
+        is_staff = request.user.is_staff
+        token = Token.objects.get(user=request.user)
+        servers = Server.objects.filter(
+            last_seen__gte=timezone.now() - datetime.timedelta(seconds=settings.HEARTBEAT_INTERVAL+5)
+        ).order_by('-last_seen')
+
+    return render(request, 'index.html', {
+        'servers': servers,
+        'token': token,
+        'is_staff': is_staff,
+        'settings': settings,
+    })