+++ django/monitor/monitor/settings.py
@@ -21,6 +21,12 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = os.environ.get('SECRET_KEY')
+DOMAIN = os.environ.get('DOMAIN')
+PORT = os.environ.get('PORT')
+WEBAUTH_USERNAME = os.environ.get('WEBAUTH_USERNAME')
+WEBAUTH_PASSWORD = os.environ.get('WEBAUTH_PASSWORD')
+
+IS_SERVICE_DOWN_SCRAPE_INTERVAL = '1m'
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
@@ -32,10 +38,11 @@ ALLOWED_HOSTS = [
     'localhost',
     'django',
     'local.django.sebastienbarbier.com',
-    'status.fromedwin.com'
+    'status.fromedwin.com',
+    os.environ.get('DOMAIN'),
 ]
 
-
+DEFAULT_AUTO_FIELD = 'django.db.models.AutoField'
 # Application definition
 
 INSTALLED_APPS = [
@@ -45,11 +52,21 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'django.contrib.sites',
     'django_prometheus',
     'rest_framework',
+    'rest_framework.authtoken',
+    'allauth',
+    'allauth.account',
+    'allauth.socialaccount',
+    'allauth.socialaccount.providers.github',
+    'monitor',
     'home',
     'health',
     'alerts',
+    'applications',
+    'clients',
+    'notifications',
 ]
 
 MIDDLEWARE = [
@@ -84,7 +101,6 @@ TEMPLATES = [
 
 WSGI_APPLICATION = 'monitor.wsgi.application'
 
-
 # Database
 # https://docs.djangoproject.com/en/3.1/ref/settings/#databases
 
@@ -95,6 +111,12 @@ DATABASES = {
     }
 }
 
+AUTHENTICATION_BACKENDS = (
+    "django.contrib.auth.backends.ModelBackend",
+    "allauth.account.auth_backends.AuthenticationBackend",
+)
+LOGIN_REDIRECT_URL = '/'
+ACCOUNT_EMAIL_VERIFICATION = 'none'
 
 # Password validation
 # https://docs.djangoproject.com/en/3.1/ref/settings/#auth-password-validators
@@ -115,6 +137,16 @@ AUTH_PASSWORD_VALIDATORS = [
 ]
 
 
+SOCIALACCOUNT_PROVIDERS = {
+    'github': {
+        'SCOPE': [
+            'user',
+        ],
+    }
+}
+
+HEARTBEAT_INTERVAL = 10 # client will call server every HEARTBEAT_INTERVAL seconds.
+
 # Internationalization
 # https://docs.djangoproject.com/en/3.1/topics/i18n/
 
@@ -128,8 +160,10 @@ USE_L10N = True
 
 USE_TZ = True
 
+SITE_ID = 1
+
 
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/3.1/howto/static-files/
 
-STATIC_URL = '/static/'
+STATIC_URL = '/django/static/'