+++ stop.sh
@@ -1,0 +1,6 @@
+# If ./.env file exist, we export variables to current system to display later
+if [ -f .env ]; then
+  export $(echo $(cat .env | sed 's/#.*//g'| xargs) | envsubst)
+fi
+
+docker-compose down